iteration2

+++ testfan2/ViewModels/RoundTeamRemoveViewModel.cs
@@ -8,7 +8,7 @@ namespace testfan2.ViewModels
     public class RoundTeamRemoveViewModel
     {
         public string Message { get; set; }
-        public decimal teamTotal { get; set; }
+        public int teamTotal { get; set; }
         public int teamCount { get; set; }
         public int DeleteId { get; set; }
     }
